chore: moved `ggh4x` to Suggests
`ggh4x` is used only when side-by-side view is specified in plot-generating functions. This is not directly used in the app. If the package user wants to use SBS view, a check was added to verify if `ggh4x` is installed. If it is not, it will throw an error and prompt the user to install the needed package.